Correction risque segfault
git-svn-id: https://svn.abls-habitat.fr/repo/Watchdog@949 9f22af27-4b39-4cfe-9f85-4bed9fef4693

+++ Watchdogd/Admin/admin_modbus.c
@@ -60,7 +60,7 @@
                    module->modbus.id, module->modbus.ip, module->modbus.bit, module->modbus.actif,
                    module->started, module->modbus.watchdog,
                    module->transaction_id, module->nbr_deconnect, module->request,
-                   module->date_next_eana,
+                   (int)module->date_next_eana,
                    (int)module->date_retente
                  );
        Write_admin ( client->connexion, chaine );
